+++ sphinxsim/sph_simulation/simulation_builder/base_simulation_builder.hpp
@@ -10,67 +10,9 @@
 namespace SPH
 {
 //=================================================================================================//
-template <class MethodContainerType>
-void SimulationBuilder::buildExternalForceIfPresent(
-    SPHSimulation &sim, MethodContainerType &main_methods, SPHBody &real_body, const json &config)
-{
-    auto &config_manager = sim.getConfigManager();
-    auto &scaling_config = config_manager.getEntity<ScalingConfig>("ScalingConfig");
-    auto &initialization_pipeline = sim.getInitializationPipeline();
-
-    if (config.contains("gravity"))
-    {
-        auto &constant_gravity =
-            main_methods.template addStateDynamics<GravityForceCK<Gravity>>(
-                real_body, Gravity(scaling_config.jsonToVecd(config.at("gravity"), "Acceleration")));
-
-        initialization_pipeline.insert_hook(
-            InitializationHookPoint::InitialCondition, [&]()
-            { constant_gravity.exec(); });
-    }
-}
-//=================================================================================================//
-template <class MethodContainerType>
-void SimulationBuilder::buildInitialConditionIfPresent(
-    SPHSimulation &sim, MethodContainerType &method_container, const json &config)
-{
-    if (config.contains("initial_conditions"))
-    {
-        auto &sph_system = sim.getSPHSystem();
-        auto &config_manager = sim.getConfigManager();
-        auto &scaling_config = config_manager.getEntity<ScalingConfig>("ScalingConfig");
-        auto &initialization_pipeline = sim.getInitializationPipeline();
-
-        auto &dynamics = method_container.addParticleDynamicsGroup();
-        for (const auto &ic : config.at("initial_conditions"))
-        {
-            const std::string name = ic.at("body_name").get<std::string>();
-            auto &real_body = sph_system.getBodyByName<RealBody>(name);
-            for (const auto &assignment : ic.at("assignments"))
-            {
-                if (assignment.contains("region"))
-                {
-                    std::string region_name = assignment.at("region").get<std::string>();
-                    auto &oriented_box = config_manager.getEntity<OrientedBox>(region_name);
-                    auto &body_region = real_body.template addBodyPart<OrientedBoxByParticle>(oriented_box);
-                    dynamics.add(&addVariableAssignment(method_container, body_region, scaling_config, assignment));
-                }
-                else
-                {
-                    dynamics.add(&addVariableAssignment(method_container, real_body, scaling_config, assignment));
-                }
-            }
-        }
-
-        initialization_pipeline.insert_hook(
-            InitializationHookPoint::InitialCondition, [&]()
-            { dynamics.exec(); });
-    }
-}
-//=================================================================================================//
-template <class MethodContainerType, class IdentifierType>
+template <class IdentifierType>
 BaseDynamics<void> &SimulationBuilder::addVariableAssignment(
-    MethodContainerType &main_methods, IdentifierType &identifier,
+    MainMethods &main_methods, IdentifierType &identifier,
     const ScalingConfig &scaling_config, const json &config)
 {
     VariableConfig var_config = parseVariableConfig(config.at("variable"));
