+++ sphinxsim/sph_simulation/simulation_builder/fluid_simulation_builder.cpp
@@ -426,4 +426,43 @@ FluidSolverConfig FluidSimulationBuilder::parseFluidSolverConfig(
     return params;
 }
 //=================================================================================================//
+void FluidSimulationBuilder::buildParticleDeletionIfPresent(
+    SPHSimulation &sim, MainMethods &main_methods, RealBody &real_body)
+{
+    auto &config_manager = sim.getConfigManager();
+    StagePipeline<SimulationHookPoint> &simulation_pipeline = sim.getSimulationPipeline();
+    auto &fluid_solver_config = config_manager.getEntity<FluidSolverConfig>("FluidSolverConfig");
+    if (fluid_solver_config.particle_deletion_)
+    {
+        auto &particle_deletion = main_methods.template addStateDynamics<
+            OutflowParticleDeletion>(real_body);
+
+        simulation_pipeline.insert_hook(
+            SimulationHookPoint::ParticleDeletion, [&]()
+            { particle_deletion.exec(); });
+    }
+}
+//=================================================================================================//
+void FluidSimulationBuilder::buildParticleSortIfPresent(
+    SPHSimulation &sim, MainMethods &main_methods, RealBody &real_body)
+{
+    auto &config_manager = sim.getConfigManager();
+    auto &fluid_solver_config = config_manager.getEntity<FluidSolverConfig>("FluidSolverConfig");
+    TimeStepper &time_stepper = sim.getSPHSolver().getTimeStepper();
+
+    if (fluid_solver_config.particle_sorting_)
+    {
+        auto &particle_sort = main_methods.addSortDynamics(real_body);
+
+        auto &simulation_pipeline = sim.getSimulationPipeline();
+        simulation_pipeline.insert_hook(
+            SimulationHookPoint::ParticleSort, [&]()
+            {
+                if (time_stepper.getIterationStep() % fluid_solver_config.sort_frequency_ == 0)
+                {
+                    particle_sort.exec();
+                } });
+    }
+}
+//=================================================================================================//
 } // namespace SPH
